Move "fieldKey" modules to top-level - aligns them with how they are actually exported
