+++ src/index.ts
@@ -34,7 +34,7 @@ import * as Assay from './labkey/Assay';
 import * as Domain from './labkey/Domain';
 import * as Exp from './labkey/Exp';
 import * as Experiment from './labkey/Experiment';
-import { FieldKey } from './labkey/fieldKey/FieldKey';
+import { FieldKey } from './labkey/FieldKey';
 import * as Filter from './labkey/Filter';
 import * as List from './labkey/List';
 import { MultiRequest } from './labkey/MultiRequest';
@@ -42,9 +42,9 @@ import * as Message from './labkey/Message';
 import * as ParticipantGroup from './labkey/ParticipantGroup';
 import * as Pipeline from './labkey/Pipeline';
 import * as Query from './labkey/Query';
-import { QueryKey } from './labkey/fieldKey/QueryKey';
+import { QueryKey } from './labkey/QueryKey';
 import * as Report from './labkey/Report';
-import { SchemaKey } from './labkey/fieldKey/SchemaKey';
+import { SchemaKey } from './labkey/SchemaKey';
 import * as Security from './labkey/Security';
 import * as Specimen from './labkey/Specimen';
 import * as Utils from './labkey/Utils';
